refactor(server): unify GitHub auto-fix into routines, drop webhook/recurring legacy
Fold GitHub auto-fix into the routines system as a hidden "managed" routine
(github triggers + comment_issue action gated by the per-issue
github_auto_fix_enabled flag). The managed routine reuses the Handler's
existing bot-comment + agent re-engagement pipeline on the github/api/manual
ingest paths.

- Add migration 067: routine.managed column + DROP webhook/webhook_action/
  webhook_event_log/recurring_issue_template/recurring_template_subscriber.
- Provision the managed routine on GitHubConnect (and remove it on disconnect)
  plus an idempotent startup reconciler for already-connected workspaces.
- Simplify ReceiveGitHubEvent to a single routine-trigger ingest path.
- Reject mutations on managed routines (403).
- Delete legacy webhook CRUD/ingest handlers and queries; drop /api/webhooks.
